Cleaned up code and added better error handling and responses for invalid tables
