+++ src/global.ts
@@ -1,0 +1,14 @@
+import { ClientDB } from "./client_db";
+import { SpacetimeDBClient } from "./spacetimedb";
+
+export type SpacetimeDBGlobals = {
+  clientDB: ClientDB;
+  spacetimeDBClient: SpacetimeDBClient | undefined;
+};
+
+declare global {
+  interface Window {
+    __SPACETIMEDB__: SpacetimeDBGlobals;
+  }
+  var __SPACETIMEDB__: SpacetimeDBGlobals;
+}
