+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from "./spacetimedb";
 export * from "./identity";
 export * from "./address";
+export * from "./client_db";
+export * from "./message_types";
